Move shim into its own file
